+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/mod.rs
@@ -6,7 +6,7 @@ use crate::runtime_extensions::{
     common::get_relocated_vm_trace,
     forge_runtime_extension::cheatcodes::{
         CheatcodeError,
-        declare::declare,
+        declare::{declare, declare_from_file},
         generate_random_felt::generate_random_felt,
         get_class_hash::get_class_hash,
         l1_handler_execute::l1_handler_execute,
@@ -185,6 +185,22 @@ impl<'a> ExtensionLogic for ForgeExtension<'a> {
 
                 handle_declare_result(declare(*state, &contract_identifier, self.contracts_data))
             }
+            "declare_from_file" => {
+                let state = &mut extended_runtime
+                    .extended_runtime
+                    .extended_runtime
+                    .hint_handler
+                    .base
+                    .state;
+
+                let sierra_path = input_reader.read::<ByteArray>()?.to_string();
+
+                handle_declare_result(declare_from_file(
+                    *state,
+                    std::path::Path::new(&sierra_path),
+                    self.contracts_data,
+                ))
+            }
             // Internal cheatcode used to pass a contract address when calling `deploy_at`.
             "set_deploy_at_address" => {
                 let contract_address = input_reader.read()?;
@@ -274,9 +290,9 @@ impl<'a> ExtensionLogic for ForgeExtension<'a> {
                     Err(CallFailure::Recoverable { panic_data }) => Ok(
                         CheatcodeHandlingResult::from_serializable(Err::<(), _>(panic_data)),
                     ),
-                    Err(CallFailure::Unrecoverable { msg }) => Err(EnhancedHintError::from(
-                        HintError::CustomHint(Box::from(msg.to_string())),
-                    )),
+                    Err(CallFailure::Unrecoverable(error)) => {
+                        Err(EnhancedHintError::from(error.into_unannotated()))
+                    }
                 }
             }
             "read_txt" => {