+++ crates/forge/tests/e2e/mod.rs
@@ -18,6 +18,7 @@ mod coverage;
 mod debugger;
 #[cfg(not(feature = "cairo-native"))]
 mod debugging;
+mod declare_from_file;
 mod declare_macro;
 mod docs_snippets_validation;
 mod enable_gas_validation;